Add __asonPut and other methods to InternalTransformInterface
The methods __asonAlignofValueofParameter() and __asonLength() have been
introduced to support serializing Object, which was problematic due to
being the superclass of Array and ArrayBufferView and therefore causing
previously static instanceof checks to become dynamic. Moreover, the
existing __asonPut() method has been added to the interface to allow it
to be called virtually; otherwise, serializing classes casted to Objects
would not include their fields.