Add a __asonNameof method to every class (#33)
* Add a __asonNameof method to every class

__asonNameof() is a virtual method that simply returns nameof<this>().
This is useful for debugging objects whose types are only known at
runtime, i.e. objects in ASON deserialization.

* Implement __asonNameof() on pre-resolved builtins

This change allows the names of String and ArrayBuffer to be logged.

* Rename InternalNameofInterface to InternalTransformInterface

There is a high chance this interface will be used for defining methods
that aren't __asonNameof. Therefore, the interface shouldn't have a name
that implies otherwise.

* Add Object to the array of pre-resolved classes

* Add a test that ensures Object can be deserialized

In the (hopefully near) future, serializing Object should be supported.

* Move the internal interface name into a constant

Moving the name into a constant, in addition to moving a check that
prevents the internal interface declaration from being processed, into
the main file keeps the corresponding logic organized.

* Add __asonPut and other methods to InternalTransformInterface

The methods __asonAlignofValueofParameter() and __asonLength() have been
introduced to support serializing Object, which was problematic due to
being the superclass of Array and ArrayBufferView and therefore causing
previously static instanceof checks to become dynamic. Moreover, the
existing __asonPut() method has been added to the interface to allow it
to be called virtually; otherwise, serializing classes casted to Objects
would not include their fields.

* Actually support serializing with Object

The previous commit laid out the foundation for serializing with Object.
This commit uses them in practice in response to the compiler adding a
few dynamic instanceof checks. A new test has been added to demonstrate
this new feature. Hypothetically, these changes allow the serialization
of interfaces as well.

* Fix a mispelling of "function" in a test name

* Replace classDeclaration.range with classDeclaration.name.range

This change leads to better error messages. Previously, errors displayed
a decorator for classes with a standalone decorator on the first line.